BC-591 move infra services to shared infra directory
