+++ apps/server/src/modules/files/repo/storageprovider.repo.spec.ts
@@ -1,6 +1,6 @@
 import { EntityManager } from '@mikro-orm/mongodb';
 import { Test, TestingModule } from '@nestjs/testing';
-import { MongoMemoryDatabaseModule } from '@src/modules/database';
+import { MongoMemoryDatabaseModule } from '@shared/testing/database';
 
 import { storageProviderFactory } from '@shared/testing';
 import { Platform } from '@mikro-orm/core';
