+++ apps/server/src/modules/management/management.module.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { FileSystemModule } from '@shared/infra/file-system';
-import { MongoMemoryDatabaseModule } from '@src/modules/database';
+import { MongoMemoryDatabaseModule } from '@shared/testing/database';
 import { DatabaseManagementController } from './controller/database-management.controller';
 import { ManagementModule } from './management.module';
 import { DatabaseManagementUc } from './uc/database-management.uc';
