+++ apps/server/src/modules/task/repo/submission.repo.integration.spec.ts
@@ -2,7 +2,7 @@ import { EntityManager } from '@mikro-orm/mongodb';
 import { Test, TestingModule } from '@nestjs/testing';
 
 import { CourseGroup } from '@shared/domain';
-import { MongoMemoryDatabaseModule } from '@src/modules/database';
+import { MongoMemoryDatabaseModule } from '@shared/testing/database';
 
 import { userFactory, courseFactory, taskFactory, submissionFactory, cleanUpCollections } from '@shared/testing';
