+++ apps/server/src/modules/user/repo/role.repo.integration.spec.ts
@@ -2,7 +2,7 @@ import { NotFoundError, ValidationError } from '@mikro-orm/core';
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
 import { Test, TestingModule } from '@nestjs/testing';
 import { Role } from '@shared/domain';
-import { MongoMemoryDatabaseModule } from '@src/modules/database';
+import { MongoMemoryDatabaseModule } from '@shared/testing/database';
 import { RoleRepo } from './role.repo';
 
 describe('role repo', () => {
