+++ apps/server/src/modules/user/uc/role.uc.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { Role } from '@shared/domain';
-import { MongoMemoryDatabaseModule } from '@src/modules/database';
+import { MongoMemoryDatabaseModule } from '@shared/testing/database';
 import { RoleRepo } from '../repo';
 import { RoleUC } from './role.uc';
