+++ apps/server/src/server.module.ts
@@ -11,7 +11,7 @@ import { CoreModule } from './core/core.module';
 import { TaskModule } from './modules/task/task.module';
 import { UserModule } from './modules/user/user.module';
 import { NewsModule } from './modules/news/news.module';
-import { MailModule } from './modules/mail/mail.module';
+import { MailModule } from './shared/infra/mail/mail.module';
 import { FilesModule } from './modules/files/files.module';
 
 import { LearnroomModule } from './modules/learnroom/learnroom.module';
