+++ apps/server/src/shared/infra/database/management/database-management.module.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { MongoMemoryDatabaseModule } from '../../../../modules/database';
+import { MongoMemoryDatabaseModule } from '../../../testing/database';
 import { DatabaseManagementModule } from './database-management.module';
 import { DatabaseManagementService } from './database-management.service';
