+++ apps/server/src/shared/infra/database/management/database-management.service.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ObjectId } from 'mongodb';
-import { MongoMemoryDatabaseModule } from '@src/modules/database';
+import { MongoMemoryDatabaseModule } from '@shared/testing/database';
 import { DatabaseManagementService } from './database-management.service';
 
 const randomChars = () => {
