+++ apps/server/src/shared/infra/mail/mail.interface.ts
